requirements.txt
PyYAML
pexpect
pygatt
pandas
numpy
